Extract processor, notification and processor api to api module - #101
