+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/FileProcessListener.kt
@@ -1,0 +1,13 @@
+package io.gitlab.arturbosch.detekt.api
+
+import org.jetbrains.kotlin.psi.KtFile
+
+/**
+ * @author Artur Bosch
+ */
+@Suppress("EmptyFunctionBlock")
+interface FileProcessListener {
+	fun onStart(files: List<KtFile>) {}
+	fun onProcess(file: KtFile) {}
+	fun onFinish(files: List<KtFile>, result: Detektion) {}
+}
