+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/Notification.kt
@@ -1,0 +1,8 @@
+package io.gitlab.arturbosch.detekt.api
+
+/**
+ * @author Artur Bosch
+ */
+interface Notification {
+	val message: String
+}
