+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/DetektProgressListener.kt
@@ -1,13 +1,14 @@
 package io.gitlab.arturbosch.detekt.cli
 
-import io.gitlab.arturbosch.detekt.core.Detektion
-import io.gitlab.arturbosch.detekt.core.FileProcessListener
+import io.gitlab.arturbosch.detekt.api.Detektion
+import io.gitlab.arturbosch.detekt.api.FileProcessListener
 import org.jetbrains.kotlin.psi.KtFile
 
 /**
  * @author Artur Bosch
  */
 class DetektProgressListener : FileProcessListener {
+
 	override fun onStart(files: List<KtFile>) {
 		kotlin.io.print("Analyzing ${files.size} kotlin files: ")
 	}
