+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/Detektion.kt
@@ -1,20 +1,14 @@
 package io.gitlab.arturbosch.detekt.core
 
+import io.gitlab.arturbosch.detekt.api.Detektion
 import io.gitlab.arturbosch.detekt.api.Finding
+import io.gitlab.arturbosch.detekt.api.Notification
 import org.jetbrains.kotlin.com.intellij.openapi.util.Key
 import org.jetbrains.kotlin.com.intellij.util.keyFMap.KeyFMap
 
 /**
  * @author Artur Bosch
  */
-interface Detektion {
-	val findings: Map<String, List<Finding>>
-	val notifications: List<Notification>
-
-	fun <V> getData(key: Key<V>): V?
-	fun <V> addData(key: Key<V>, value: V)
-}
-
 data class DetektResult(override val findings: Map<String, List<Finding>>,
 						override val notifications: List<Notification>) : Detektion {
