+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/Notifications.kt
@@ -1,14 +1,11 @@
 package io.gitlab.arturbosch.detekt.core
 
+import io.gitlab.arturbosch.detekt.api.Notification
 import java.nio.file.Path
 
 /**
  * @author Artur Bosch
  */
-interface Notification {
-	val message: String
-}
-
 class ModificationNotification(path: Path) : Notification {
 	override val message: String = "File $path was modified."
